Refactored Dashboard Screen to be integrated into the react project and can be nagivated to with localhost:3000/dashboard
