+++ frontend/sign-on-form/src/screens/DashboardScreen.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Dashboard from '../components/Dashboard/Dashboard.jsx';
+
+export default function DashboardScreen() {
+  return (
+    <div>
+      <Dashboard/>
+    </div>
+  )
+}
